+++ lib/sucker_punch/testing/inline.rb
@@ -1,11 +1,37 @@
 require 'sucker_punch'
-require 'celluloid/proxy/abstract'
-require 'celluloid/proxy/sync'
-require 'celluloid/proxy/actor'
 
-class Celluloid::Proxy::Cell < Celluloid::Proxy::Sync
-  def async(method_name = nil, *args, &block)
-    self
+# Include this in your tests to simulate
+# immediate execution of your asynchronous jobs
+#
+# class LogJob
+#   include SuckerPunch::Job
+#
+#   def perform(*args)
+#     # log the things
+#   end
+# end
+#
+# To trigger asynchronous job:
+#
+#   LogJob.perform_async(1, 2, 3)
+#
+# Include inline testing lib:
+#
+# require 'sucker_punch/testing/inline"
+#
+# LogJob.perform_async(1, 2, 3) is now synchronous
+# LogJob.perform_in(1, 2, 3) is now synchronous
+#
+module SuckerPunch
+  module Job
+    module ClassMethods
+      def perform_async(*args)
+        self.new.perform(*args)
+      end
+
+      def perform_in(_, *args)
+        self.new.perform(*args)
+      end
+    end
   end
 end
-