+++ spec/spec_helper.rb
@@ -1,9 +1,0 @@
-begin
-  require 'pry'
-rescue LoadError
-end
-
-require 'sucker_punch'
-
-RSpec.configure do |config|
-end