+++ test/test_helper.rb
@@ -1,0 +1,8 @@
+$LOAD_PATH.unshift File.expand_path('../../lib', __FILE__)
+require 'sucker_punch'
+
+require 'pry'
+require 'minitest/autorun'
+require 'minitest/pride'
+
+SuckerPunch.logger = nil